sparc: fix kc compilation, wire it into tests/c/mini
kc had two real compile/link bugs (align()/maxround() return-type
mismatch vs cc.h, EXTERN vs extern on hintabsize causing a duplicate
definition), plus mkfiles/sparc/mkfile was an unedited copy-paste of
power's mkfile. Fixed all three, then brought kc up through
tests/c/mini: wired .k object support into lib_toolchain/libmach's ar
dispatch table (kobj.c existed but was never registered), and found
two sparc-specific runtime bugs where this backend's REGSP (%g1) has
no relation to real hardware registers -- the kernel never
initializes it, and it collides with the Linux syscall-number
register. Full root-cause writeup in
docs/claude_notes/notes_arch_sparc.txt.

+++ compilers/kc/swt.c
@@ -505,10 +505,10 @@ zaddr(Biobuf *b, Adr *a, int s)
 	}
 }
 
-long
-align(long i, Type *t, int op)
+int32
+align(int32 i, Type *t, int op)
 {
-	long o;
+	int32 o;
 	Type *v;
 	int w;
 
@@ -572,8 +572,8 @@ align(long i, Type *t, int op)
 	return o;
 }
 
-long
-maxround(long max, long v)
+int32
+maxround(int32 max, int32 v)
 {
 	v = round(v, SZ_LONG);
 	if(v > max)
